requirements.txt
requests
python-whois
beautifulsoup4
bs4
